requirements.txt
faiss-cpu

pdfplumber 

langchain-core
langchain>=0.2.10
langchain-community>=0.2.10
langchain_text_splitters

unstructured

unstructured[all-docs]
fastembed
pikepdf
sentence-transformers
elevenlabs

fastapi
uvicorn
python-dotenv
together
